Move dialogues into data folder
