+++ data/dialogues/coren_dialogue.js
