+++ data/dialogues/thalos_dialogue.js
