+++ scripts/npc/arvalin.js
@@ -1,5 +1,5 @@
 import { startDialogueTree } from '../dialogueSystem.js';
-import { createArvalinDialogue } from '../npc_dialogues/arvalin_dialogue.js';
+import { createArvalinDialogue } from '../../data/dialogues/arvalin_dialogue.js';
 
 export async function interact() {
   const dialogue = await createArvalinDialogue();
