+++ scripts/npc/caelen.js
@@ -1,5 +1,5 @@
 import { startDialogueTree } from '../dialogueSystem.js';
-import { caelenDialogue } from '../npc_dialogues/caelen_dialogue.js';
+import { caelenDialogue } from '../../data/dialogues/caelen_dialogue.js';
 
 export function interact() {
   startDialogueTree(caelenDialogue);
