+++ scripts/npc/coren.js
@@ -1,5 +1,5 @@
 import { startDialogueTree, showDialogue } from '../dialogueSystem.js';
-import { corenDialogue } from '../npc_dialogues/coren_dialogue.js';
+import { corenDialogue } from '../../data/dialogues/coren_dialogue.js';
 
 export function interact() {
   showDialogue('Coren', () => startDialogueTree(corenDialogue));
