+++ scripts/npc/darius_the_conversationalist.js
@@ -1,5 +1,5 @@
 import { startDialogueTree } from '../dialogueSystem.js';
-import { createDariusDialogue } from '../npc_dialogues/darius_the_conversationalist.js';
+import { createDariusDialogue } from '../../data/dialogues/darius_the_conversationalist.js';
 
 export async function interact() {
   const dialogue = await createDariusDialogue();
