+++ scripts/npc/first_memory.js
@@ -1,5 +1,5 @@
 import { startDialogueTree, showDialogue } from '../dialogueSystem.js';
-import { firstMemoryDialogue } from '../npc_dialogues/first_memory_dialogue.js';
+import { firstMemoryDialogue } from '../../data/dialogues/first_memory_dialogue.js';
 import { npcAppearance } from '../npc_data.js';
 
 export function interact() {
