+++ scripts/npc/flamebound.js
@@ -1,5 +1,5 @@
 import { startDialogueTree } from '../dialogueSystem.js';
-import { flameboundDialogue } from '../npc_dialogues/flamebound_dialogue.js';
+import { flameboundDialogue } from '../../data/dialogues/flamebound_dialogue.js';
 
 export function interact() {
   startDialogueTree(flameboundDialogue);
