+++ scripts/npc/forge_npc.js
@@ -1,5 +1,5 @@
 import { startDialogueTree } from '../dialogueSystem.js';
-import { createForgeDialogue } from '../npc_dialogues/forge_npc.js';
+import { createForgeDialogue } from '../../data/dialogues/forge_npc.js';
 
 export async function interact() {
   const dialogue = await createForgeDialogue();
