+++ scripts/npc/fork_guide.js
@@ -1,5 +1,5 @@
 import { startDialogueTree } from '../dialogueSystem.js';
-import { forkGuideDialogue } from '../npc_dialogues/fork_guide_dialogue.js';
+import { forkGuideDialogue } from '../../data/dialogues/fork_guide_dialogue.js';
 
 export function interact() {
   startDialogueTree(forkGuideDialogue);
