+++ scripts/npc/goblin_quest_giver.js
@@ -1,5 +1,5 @@
 import { startDialogueTree } from '../dialogueSystem.js';
-import { goblinQuestDialogue } from '../npc_dialogues/goblin_quest_giver.js';
+import { goblinQuestDialogue } from '../../data/dialogues/goblin_quest_giver.js';
 
 export function interact() {
   startDialogueTree(goblinQuestDialogue);
