+++ scripts/npc/kaelor_the_weaver.js
@@ -1,5 +1,5 @@
 import { startDialogueTree } from '../dialogueSystem.js';
-import { createKaelorDialogue } from '../npc_dialogues/kaelor_the_weaver.js';
+import { createKaelorDialogue } from '../../data/dialogues/kaelor_the_weaver.js';
 
 export async function interact() {
   const dialogue = await createKaelorDialogue();
