+++ scripts/npc/myralen.js
@@ -1,5 +1,5 @@
 import { startDialogueTree } from '../dialogueSystem.js';
-import { createMyralenDialogue } from '../npc_dialogues/myralen_dialogue.js';
+import { createMyralenDialogue } from '../../data/dialogues/myralen_dialogue.js';
 
 export async function interact() {
   const dialogue = await createMyralenDialogue();
