+++ scripts/npc/mystic_orb.js
@@ -1,5 +1,5 @@
 import { startDialogueTree } from '../dialogueSystem.js';
-import { mysticOrbDialogue } from '../npc_dialogues/mystic_orb_dialogue.js';
+import { mysticOrbDialogue } from '../../data/dialogues/mystic_orb_dialogue.js';
 
 export function interact() {
   startDialogueTree(mysticOrbDialogue);
