+++ scripts/npc/shade_sage.js
@@ -1,5 +1,5 @@
 import { startDialogueTree } from '../dialogueSystem.js';
-import { shadeSageDialogue } from '../npc_dialogues/shade_sage_dialogue.js';
+import { shadeSageDialogue } from '../../data/dialogues/shade_sage_dialogue.js';
 
 export function interact() {
   startDialogueTree(shadeSageDialogue);
