+++ scripts/npc/silari_the_analyst.js
@@ -1,5 +1,5 @@
 import { startDialogueTree } from '../dialogueSystem.js';
-import { createSilariDialogue } from '../npc_dialogues/silari_the_analyst.js';
+import { createSilariDialogue } from '../../data/dialogues/silari_the_analyst.js';
 
 export async function interact() {
   const dialogue = await createSilariDialogue();
