+++ scripts/npc/syranel.js
@@ -1,5 +1,5 @@
 import { startDialogueTree } from '../dialogueSystem.js';
-import { syranelDialogue } from '../npc_dialogues/syranel_dialogue.js';
+import { syranelDialogue } from '../../data/dialogues/syranel_dialogue.js';
 
 export function interact() {
   startDialogueTree(syranelDialogue);
