+++ scripts/npc/thalos.js
@@ -1,5 +1,5 @@
 import { startDialogueTree } from '../dialogueSystem.js';
-import { thalosDialogue } from '../npc_dialogues/thalos_dialogue.js';
+import { thalosDialogue } from '../../data/dialogues/thalos_dialogue.js';
 
 export function interact() {
   startDialogueTree(thalosDialogue);
