+++ scripts/npc/vaelin.js
@@ -1,5 +1,5 @@
 import { startDialogueTree } from '../dialogueSystem.js';
-import { vaelinDialogue } from '../npc_dialogues/vaelin_dialogue.js';
+import { vaelinDialogue } from '../../data/dialogues/vaelin_dialogue.js';
 
 export function interact() {
   startDialogueTree(vaelinDialogue);
