+++ scripts/npc/verrel_the_voice.js
@@ -1,5 +1,5 @@
 import { startDialogueTree } from '../dialogueSystem.js';
-import { createVerrelDialogue } from '../npc_dialogues/verrel_the_voice.js';
+import { createVerrelDialogue } from '../../data/dialogues/verrel_the_voice.js';
 
 export async function interact() {
   const dialogue = await createVerrelDialogue();
