+++ scripts/npc/watcher.js
@@ -1,5 +1,5 @@
 import { startDialogueTree } from '../dialogueSystem.js';
-import { watcherDialogue } from '../npc_dialogues/watcher_dialogue.js';
+import { watcherDialogue } from '../../data/dialogues/watcher_dialogue.js';
 
 export function interact() {
   startDialogueTree(watcherDialogue);
