--- scripts/npc_dialogues/watcher_dialogue.js
@@ -1,8 +1,0 @@
-export const watcherDialogue = [
-  {
-    text: 'You sense eyes upon you though none are seen.',
-    options: [
-      { label: 'Move quietly onward.', goto: null, onChoose: () => import('../dialogue_state.js').then(m => m.discoverLore('shadow_fork')) }
-    ]
-  }
-];
